[REMANIEMENT] Extrais <KitDeTeleversement />
… avant de le réutiliser sur le téléversement des modèles de mesure.
Ce commit conserve le comportement existant sur le téléversement de services, mais avec le nouveau composant.
