+++ svelte/lib/ui/televersement/KitDeTeleversement.types.ts
@@ -1,0 +1,9 @@
+export type EtatTeleversement =
+  | 'EnAttente'
+  | 'EnCoursEnvoi'
+  | 'Invalide'
+  | 'Valide';
+
+export enum FormatAccepte {
+  Excel = 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
+}
